Update cmake list and refactor resource usage
The cmake list has been updated to include 'Deploy.cmake' and to set an automatic path prefix for moc. The application now uses Qt resources system to handle and deploy static resources such as images, significantly simplifying the way resources are referenced in the code. This change also includes some refactoring of how include directories are managed in the project.